Avoid making a new module for this function
